+++ src/assets/AssetsDetectionController.test.ts
@@ -1,15 +1,21 @@
 import { createSandbox, SinonStub, stub } from 'sinon';
 import nock from 'nock';
 import { BN } from 'ethereumjs-util';
-import contractMap from '@metamask/contract-metadata';
 import {
   NetworkController,
   NetworksChainId,
 } from '../network/NetworkController';
 import { PreferencesController } from '../user/PreferencesController';
-import { AssetsController } from './AssetsController';
+import { ControllerMessenger } from '../ControllerMessenger';
+import { TokensController } from './TokensController';
+import { CollectiblesController } from './CollectiblesController';
 import { AssetsContractController } from './AssetsContractController';
 import { AssetsDetectionController } from './AssetsDetectionController';
+import {
+  TokenListController,
+  GetTokenListState,
+  TokenListStateChange,
+} from './TokenListController';
 
 const DEFAULT_INTERVAL = 180000;
 const MAINNET = 'mainnet';
@@ -17,12 +23,90 @@ const ROPSTEN = 'ropsten';
 const TOKENS = [{ address: '0xfoO', symbol: 'bar', decimals: 2 }];
 const OPEN_SEA_HOST = 'https://api.opensea.io';
 const OPEN_SEA_PATH = '/api/v1';
+const TOKEN_END_POINT_API = 'https://token-api.airswap-prod.codefi.network';
+const sampleTokenList = [
+  {
+    address: '0x514910771af9ca656af840dff83e8264ecf986ca',
+    symbol: 'LINK',
+    decimals: 18,
+    occurrences: 11,
+    aggregators: [
+      'paraswap',
+      'pmm',
+      'airswapLight',
+      'zeroEx',
+      'bancor',
+      'coinGecko',
+      'zapper',
+      'kleros',
+      'zerion',
+      'cmc',
+      'oneInch',
+    ],
+    name: 'Chainlink',
+  },
+  {
+    address: '0x1f573d6fb3f13d689ff844b4ce37794d79a7ff1c',
+    symbol: 'BNT',
+    decimals: 18,
+    occurrences: 11,
+    aggregators: [
+      'paraswap',
+      'pmm',
+      'airswapLight',
+      'zeroEx',
+      'bancor',
+      'coinGecko',
+      'zapper',
+      'kleros',
+      'zerion',
+      'cmc',
+      'oneInch',
+    ],
+    name: 'Bancor',
+  },
+  {
+    address: '0x6810e776880c02933d47db1b9fc05908e5386b96',
+    symbol: 'GNO',
+    name: 'Gnosis',
+    decimals: 18,
+    occurrences: 10,
+    aggregators: [
+      'paraswap',
+      'airswapLight',
+      'zeroEx',
+      'bancor',
+      'coinGecko',
+      'zapper',
+      'kleros',
+      'zerion',
+      'cmc',
+      'oneInch',
+    ],
+  },
+];
+function getTokenListMessenger() {
+  const controllerMessenger = new ControllerMessenger<
+    GetTokenListState,
+    TokenListStateChange
+  >();
+  const messenger = controllerMessenger.getRestricted<
+    'TokenListController',
+    never,
+    never
+  >({
+    name: 'TokenListController',
+  });
+  return messenger;
+}
 
 describe('AssetsDetectionController', () => {
   let assetsDetection: AssetsDetectionController;
   let preferences: PreferencesController;
   let network: NetworkController;
-  let assets: AssetsController;
+  let tokensController: TokensController;
+  let collectiblesController: CollectiblesController;
+  let tokenList: TokenListController;
   let assetsContract: AssetsContractController;
   let getBalancesInSingleCall: SinonStub<
     Parameters<AssetsContractController['getBalancesInSingleCall']>,
@@ -30,11 +114,15 @@ describe('AssetsDetectionController', () => {
   >;
   const sandbox = createSandbox();
 
-  beforeEach(() => {
+  beforeEach(async () => {
     preferences = new PreferencesController();
     network = new NetworkController();
     assetsContract = new AssetsContractController();
-    assets = new AssetsController({
+    tokensController = new TokensController({
+      onPreferencesStateChange: (listener) => preferences.subscribe(listener),
+      onNetworkStateChange: (listener) => network.subscribe(listener),
+    });
+    collectiblesController = new CollectiblesController({
       onPreferencesStateChange: (listener) => preferences.subscribe(listener),
       onNetworkStateChange: (listener) => network.subscribe(listener),
       getAssetName: assetsContract.getAssetName.bind(assetsContract),
@@ -43,16 +131,35 @@ describe('AssetsDetectionController', () => {
         assetsContract,
       ),
     });
+    nock(TOKEN_END_POINT_API)
+      .get(`/tokens/${NetworksChainId.mainnet}`)
+      .reply(200, sampleTokenList)
+      .persist();
+    const messenger = getTokenListMessenger();
+    tokenList = new TokenListController({
+      chainId: NetworksChainId.mainnet,
+      useStaticTokenList: false,
+      onNetworkStateChange: (listener) => network.subscribe(listener),
+      onPreferencesStateChange: (listener) => preferences.subscribe(listener),
+      messenger,
+    });
+    await tokenList.start();
     getBalancesInSingleCall = sandbox.stub();
     assetsDetection = new AssetsDetectionController({
-      onAssetsStateChange: (listener) => assets.subscribe(listener),
+      onTokensStateChange: (listener) => tokensController.subscribe(listener),
+      onCollectiblesStateChange: (listener) =>
+        collectiblesController.subscribe(listener),
       onPreferencesStateChange: (listener) => preferences.subscribe(listener),
       onNetworkStateChange: (listener) => network.subscribe(listener),
-      getOpenSeaApiKey: () => assets.openSeaApiKey,
+      getOpenSeaApiKey: () => collectiblesController.openSeaApiKey,
       getBalancesInSingleCall: (getBalancesInSingleCall as unknown) as AssetsContractController['getBalancesInSingleCall'],
-      addTokens: assets.addTokens.bind(assets),
-      addCollectible: assets.addCollectible.bind(assets),
-      getAssetsState: () => assets.state,
+      addTokens: tokensController.addTokens.bind(tokensController),
+      addCollectible: collectiblesController.addCollectible.bind(
+        collectiblesController,
+      ),
+      getTokensState: () => tokensController.state,
+      getTokenListState: () => tokenList.state,
+      getCollectiblesState: () => collectiblesController.state,
     });
 
     nock(OPEN_SEA_HOST)
@@ -153,6 +260,7 @@ describe('AssetsDetectionController', () => {
   afterEach(() => {
     nock.cleanAll();
     sandbox.reset();
+    tokenList.destroy();
   });
 
   it('should set default config', () => {
@@ -176,17 +284,24 @@ describe('AssetsDetectionController', () => {
       );
       new AssetsDetectionController(
         {
-          onAssetsStateChange: (listener) => assets.subscribe(listener),
+          onTokensStateChange: (listener) =>
+            tokensController.subscribe(listener),
+          onCollectiblesStateChange: (listener) =>
+            collectiblesController.subscribe(listener),
           onPreferencesStateChange: (listener) =>
             preferences.subscribe(listener),
           onNetworkStateChange: (listener) => network.subscribe(listener),
-          getOpenSeaApiKey: () => assets.openSeaApiKey,
+          getOpenSeaApiKey: () => collectiblesController.openSeaApiKey,
           getBalancesInSingleCall: assetsContract.getBalancesInSingleCall.bind(
             assetsContract,
           ),
-          addTokens: assets.addTokens.bind(assets),
-          addCollectible: assets.addCollectible.bind(assets),
-          getAssetsState: () => assets.state,
+          addTokens: tokensController.addTokens.bind(tokensController),
+          addCollectible: collectiblesController.addCollectible.bind(
+            collectiblesController,
+          ),
+          getTokensState: () => tokensController.state,
+          getTokenListState: () => tokenList.state,
+          getCollectiblesState: () => collectiblesController.state,
         },
         { interval: 10 },
       );
@@ -221,17 +336,24 @@ describe('AssetsDetectionController', () => {
       );
       new AssetsDetectionController(
         {
-          onAssetsStateChange: (listener) => assets.subscribe(listener),
+          onTokensStateChange: (listener) =>
+            tokensController.subscribe(listener),
+          onCollectiblesStateChange: (listener) =>
+            collectiblesController.subscribe(listener),
           onPreferencesStateChange: (listener) =>
             preferences.subscribe(listener),
           onNetworkStateChange: (listener) => network.subscribe(listener),
-          getOpenSeaApiKey: () => assets.openSeaApiKey,
+          getOpenSeaApiKey: () => collectiblesController.openSeaApiKey,
           getBalancesInSingleCall: assetsContract.getBalancesInSingleCall.bind(
             assetsContract,
           ),
-          addTokens: assets.addTokens.bind(assets),
-          addCollectible: assets.addCollectible.bind(assets),
-          getAssetsState: () => assets.state,
+          addTokens: tokensController.addTokens.bind(tokensController),
+          addCollectible: collectiblesController.addCollectible.bind(
+            collectiblesController,
+          ),
+          getTokensState: () => tokensController.state,
+          getTokenListState: () => tokenList.state,
+          getCollectiblesState: () => collectiblesController.state,
         },
         { interval: 10, networkType: ROPSTEN },
       );
@@ -246,7 +368,7 @@ describe('AssetsDetectionController', () => {
   it('should detect and add collectibles correctly', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibles).toStrictEqual([
+    expect(collectiblesController.state.collectibles).toStrictEqual([
       {
         address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
         description: 'Description 2574',
@@ -259,7 +381,7 @@ describe('AssetsDetectionController', () => {
 
   it('should detect, add collectibles and do nor remove not detected collectibles correctly', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
-    await assets.addCollectible(
+    await collectiblesController.addCollectible(
       '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
       2573,
       {
@@ -269,7 +391,7 @@ describe('AssetsDetectionController', () => {
       },
     );
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibles).toStrictEqual([
+    expect(collectiblesController.state.collectibles).toStrictEqual([
       {
         address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
         description: 'Description 2573',
@@ -290,21 +412,21 @@ describe('AssetsDetectionController', () => {
   it('should not autodetect collectibles that exist in the ignoreList', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x2' });
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibles).toHaveLength(1);
-    expect(assets.state.ignoredCollectibles).toHaveLength(0);
-    assets.removeAndIgnoreCollectible(
+    expect(collectiblesController.state.collectibles).toHaveLength(1);
+    expect(collectiblesController.state.ignoredCollectibles).toHaveLength(0);
+    collectiblesController.removeAndIgnoreCollectible(
       '0x1d963688fe2209a98db35c67a041524822cf04ff',
       2577,
     );
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibles).toHaveLength(0);
-    expect(assets.state.ignoredCollectibles).toHaveLength(1);
+    expect(collectiblesController.state.collectibles).toHaveLength(0);
+    expect(collectiblesController.state.ignoredCollectibles).toHaveLength(1);
   });
 
   it('should not detect and add collectibles if there is no selectedAddress', async () => {
     assetsDetection.configure({ networkType: MAINNET });
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibles).toStrictEqual([]);
+    expect(collectiblesController.state.collectibles).toStrictEqual([]);
   });
 
   it('should not detect and add collectibles to the wrong selectedAddress', async () => {
@@ -312,13 +434,13 @@ describe('AssetsDetectionController', () => {
       networkType: MAINNET,
       selectedAddress: '0x9',
     });
-    assets.configure({ selectedAddress: '0x9' });
+    collectiblesController.configure({ selectedAddress: '0x9' });
     assetsDetection.detectCollectibles();
     assetsDetection.configure({ selectedAddress: '0x12' });
-    assets.configure({ selectedAddress: '0x12' });
+    collectiblesController.configure({ selectedAddress: '0x12' });
     await new Promise((res) => setTimeout(() => res(true), 1000));
     expect(assetsDetection.config.selectedAddress).toStrictEqual('0x12');
-    expect(assets.state.collectibles).toStrictEqual([]);
+    expect(collectiblesController.state.collectibles).toStrictEqual([]);
   });
 
   it('should not add collectible if collectible or collectible contract has no information to display', async () => {
@@ -370,8 +492,10 @@ describe('AssetsDetectionController', () => {
     assetsDetection.configure({ selectedAddress: '0x1', networkType: MAINNET });
     await assetsDetection.detectCollectibles();
     // First fetch to API, only gets information from contract ending in HH
-    expect(assets.state.collectibles).toStrictEqual([collectibleHH2574]);
-    expect(assets.state.collectibleContracts).toStrictEqual([
+    expect(collectiblesController.state.collectibles).toStrictEqual([
+      collectibleHH2574,
+    ]);
+    expect(collectiblesController.state.collectibleContracts).toStrictEqual([
       collectibleContractHH,
     ]);
     // During next call of assets detection, API succeds returning contract ending in gg information
@@ -432,12 +556,12 @@ describe('AssetsDetectionController', () => {
 
     // Now user should have respective collectibles
     await assetsDetection.detectCollectibles();
-    expect(assets.state.collectibleContracts).toStrictEqual([
+    expect(collectiblesController.state.collectibleContracts).toStrictEqual([
       collectibleContractHH,
       collectibleContractII,
       collectibleContractGG,
     ]);
-    expect(assets.state.collectibles).toStrictEqual([
+    expect(collectiblesController.state.collectibles).toStrictEqual([
       collectibleHH2574,
       collectibleII2577,
       collectibleGG2574,
@@ -447,26 +571,25 @@ describe('AssetsDetectionController', () => {
   it('should detect tokens correctly', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
     getBalancesInSingleCall.resolves({
-      '0x6810e776880C02933D47DB1b9fc05908e5386b96': new BN(1),
+      '0x6810e776880c02933d47db1b9fc05908e5386b96': new BN(1),
     });
     await assetsDetection.detectTokens();
-    expect(assets.state.tokens).toStrictEqual([
+    expect(tokensController.state.tokens).toStrictEqual([
       {
         address: '0x6810e776880C02933D47DB1b9fc05908e5386b96',
+        symbol: 'GNO',
         decimals: 18,
         image: undefined,
-        symbol: 'GNO',
       },
     ]);
   });
-
   it('should update the tokens list when new tokens are detected', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
     getBalancesInSingleCall.resolves({
-      '0x6810e776880C02933D47DB1b9fc05908e5386b96': new BN(1),
+      '0x6810e776880c02933d47db1b9fc05908e5386b96': new BN(1),
     });
     await assetsDetection.detectTokens();
-    expect(assets.state.tokens).toStrictEqual([
+    expect(tokensController.state.tokens).toStrictEqual([
       {
         address: '0x6810e776880C02933D47DB1b9fc05908e5386b96',
         decimals: 18,
@@ -475,10 +598,10 @@ describe('AssetsDetectionController', () => {
       },
     ]);
     getBalancesInSingleCall.resolves({
-      '0x514910771AF9Ca656af840dff83E8264EcF986CA': new BN(1),
+      '0x514910771af9ca656af840dff83e8264ecf986ca': new BN(1),
     });
     await assetsDetection.detectTokens();
-    expect(assets.state.tokens).toStrictEqual([
+    expect(tokensController.state.tokens).toStrictEqual([
       {
         address: '0x6810e776880C02933D47DB1b9fc05908e5386b96',
         decimals: 18,
@@ -487,9 +610,9 @@ describe('AssetsDetectionController', () => {
       },
       {
         address: '0x514910771AF9Ca656af840dff83E8264EcF986CA',
+        symbol: 'LINK',
         decimals: 18,
         image: undefined,
-        symbol: 'LINK',
       },
     ]);
   });
@@ -497,50 +620,50 @@ describe('AssetsDetectionController', () => {
   it('should call getBalancesInSingle with token address that is not present on the asset state', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
     getBalancesInSingleCall.resolves({
-      '0x6810e776880C02933D47DB1b9fc05908e5386b96': new BN(1),
+      '0x6810e776880c02933d47db1b9fc05908e5386b96': new BN(1),
     });
-    const tokensToDetect: string[] = [];
-    for (const address in contractMap) {
-      const contract = contractMap[address];
-      if (contract.erc20) {
-        tokensToDetect.push(address);
-      }
-    }
+    const tokensToDetect: string[] = Object.keys(tokenList.state.tokenList);
     await assetsDetection.detectTokens();
-    expect(getBalancesInSingleCall.calledWith('0x1', tokensToDetect)).toBe(
-      true,
-    );
+    expect(
+      getBalancesInSingleCall
+        .getCall(0)
+        .calledWithExactly('0x1', tokensToDetect),
+    ).toBe(true);
     getBalancesInSingleCall.resolves({
-      '0x514910771AF9Ca656af840dff83E8264EcF986CA': new BN(1),
+      '0x514910771af9ca656af840dff83e8264ecf986ca': new BN(1),
     });
     const updatedTokensToDetect = tokensToDetect.filter(
-      (address) => address !== '0x6810e776880C02933D47DB1b9fc05908e5386b96',
+      (address) => address !== '0x6810e776880c02933d47db1b9fc05908e5386b96',
     );
     await assetsDetection.detectTokens();
     expect(
-      getBalancesInSingleCall.calledWith('0x1', updatedTokensToDetect),
+      getBalancesInSingleCall
+        .getCall(1)
+        .calledWithExactly('0x1', updatedTokensToDetect),
     ).toBe(true);
   });
 
   it('should not autodetect tokens that exist in the ignoreList', async () => {
     assetsDetection.configure({ networkType: MAINNET, selectedAddress: '0x1' });
     getBalancesInSingleCall.resolves({
-      '0x6810e776880C02933D47DB1b9fc05908e5386b96': new BN(1),
+      '0x514910771af9ca656af840dff83e8264ecf986ca': new BN(1),
     });
     await assetsDetection.detectTokens();
 
-    assets.removeAndIgnoreToken('0x6810e776880C02933D47DB1b9fc05908e5386b96');
+    tokensController.removeAndIgnoreToken(
+      '0x514910771af9ca656af840dff83e8264ecf986ca',
+    );
     await assetsDetection.detectTokens();
-    expect(assets.state.tokens).toStrictEqual([]);
+    expect(tokensController.state.tokens).toStrictEqual([]);
   });
 
   it('should not detect tokens if there is no selectedAddress set', async () => {
     assetsDetection.configure({ networkType: MAINNET });
     getBalancesInSingleCall.resolves({
-      '0x6810e776880C02933D47DB1b9fc05908e5386b96': new BN(1),
+      '0x514910771af9ca656af840dff83e8264ecf986ca': new BN(1),
     });
     await assetsDetection.detectTokens();
-    expect(assets.state.tokens).toStrictEqual([]);
+    expect(tokensController.state.tokens).toStrictEqual([]);
   });
 
   it('should subscribe to new sibling detecting assets when account changes', async () => {
@@ -569,7 +692,7 @@ describe('AssetsDetectionController', () => {
       },
     });
     expect(network.state.provider.type).toStrictEqual(firstNetworkType);
-    assets.update({ tokens: TOKENS });
+    tokensController.update({ tokens: TOKENS });
     expect(assetsDetection.config.tokens).toStrictEqual(TOKENS);
   });
 });