+++ CHANGELOG.md
@@ -6,7 +6,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [10.1.0] - 2021-06-7
+## [13.0.0] - 2021-07-12
+### Changed
+- **BREAKING**: Remove AssetsController and add CollectiblesController and TokensController in its place ([#518](https://github.com/MetaMask/controllers/pull/518))
+
+## [12.1.0] - 2021-07-09
+### Added
+- Support for custom network and gas estimation ([#505](https://github.com/MetaMask/controllers/pull/505))
+- A fallback for the token list API service based on user preference ([#517](https://github.com/MetaMask/controllers/pull/517))
+
+## [12.0.0] - 2021-07-07
+### Added
+- Add GasFeeController to provide gas fee estimates [#494](https://github.com/MetaMask/controllers/pull/494)
+
+### Changed
+- **BREAKING**: Add chainId support to TokenRatesControllers [#476](https://github.com/MetaMask/controllers/pull/476)
+  - The breaking change here is that TokenRatesController constructor now requires a onNetworkStateChange listener
+- Add iconUrl to Token type in TokenListController [#512](https://github.com/MetaMask/controllers/pull/512)
+
+## [11.0.0] - 2021-07-02
+### Changed
+- We accidentally shipped a breaking change in v10.2.0. The changelog has been updated to explain the breaking change, and it has been republished as v11.0.0.
+
+## [10.2.0] - 2021-06-30 [DEPRECATED]
+### Added
+- **BREAKING**: Add TokenListController to fetch the token list from token services API ([#478](https://github.com/MetaMask/controllers/pull/478))
+  - The breaking change here is that `AssetsDetectionController` now requires `getTokenListState` as a constructor parameter.
+- Update `@ethereumjs-tx` to `@ethereumjs/tx` and add `@ethereumjs/common` to support EIP1559 compliant transactions ([#489](https://github.com/MetaMask/controllers/pull/489))
+
+### Changed
+- Bump @metamask/contract-metadata from 1.25.0 to 1.26.0 and 1.26.0 to 1.27.0 ([#492](https://github.com/MetaMask/controllers/pull/492),[#501](https://github.com/MetaMask/controllers/pull/501))
+
+## [10.1.0] - 2021-06-07
 ### Added
 - Export BaseControllerV2 Json type ([#482](https://github.com/MetaMask/controllers/pull/482))
 
@@ -19,7 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [10.0.0]
 ### Fixed
-- **BREAKING**: Fix stale conversionRate after switching network ([#465](git+https://github.com/MetaMask/controllers/pull/465))
+- **BREAKING**: Fix stale conversionRate after switching network ([#465](https://github.com/MetaMask/controllers/pull/465))
   - The breaking change is the change in type of the `conversionRate` state of the `CurrencyRateController` - it's now nullable.
 
 ## [9.1.0] - 2021-05-20
@@ -256,10 +287,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rebrand as `@metamask/controllers` (#226)
 - Use yarn & drop `npm-shrinkwrap.json` (#193)
 
-## Removed
+### Removed
 - Remove shapeshift controller (#209)
 
-[Unreleased]: https://github.com/MetaMask/controllers/compare/v10.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/controllers/compare/v13.0.0...HEAD
+[13.0.0]: https://github.com/MetaMask/controllers/compare/v12.1.0...v13.0.0
+[12.1.0]: https://github.com/MetaMask/controllers/compare/v12.0.0...v12.1.0
+[12.0.0]: https://github.com/MetaMask/controllers/compare/v11.0.0...v12.0.0
+[11.0.0]: https://github.com/MetaMask/controllers/compare/v10.2.0...v11.0.0
+[10.2.0]: https://github.com/MetaMask/controllers/compare/v10.1.0...v10.2.0
 [10.1.0]: https://github.com/MetaMask/controllers/compare/v10.0.0...v10.1.0
 [10.0.0]: https://github.com/MetaMask/controllers/compare/v9.1.0...v10.0.0
 [9.1.0]: https://github.com/MetaMask/controllers/compare/v9.0.0...v9.1.0